feat(complete): bars_batch 连接池并行区间回补 + day 解析价格 f64 精度修复 (v1.9.0, cli v1.7.1)
- Client::bars_batch：多只股票任意周期 K 线区间批量拉取（k_batch 同款
  连接池并行；每只走 bars_range 全套防护），分钟线全市场回补配套
- file::day::Day 价格/金额/量字段 f32→f64：源为 u32 整数分，f32 仅 ~7 位
  有效数字引入尾差（实测 139,323,904 股 CSV 输出丢 .04）；CH DDL 同步
  Float64（cli v1.7.1）；compute_pre_pct 签名同步
- 快照更新为 f64 精确输出；bars_batch 与 bars_range 实盘对照测试

+++ src/file/day/fq.rs
@@ -10,12 +10,12 @@ use std::path::Path;
 pub struct Day {
     pub date: String,
     pub code: String,
-    pub open: f32,
-    pub high: f32,
-    pub low: f32,
-    pub close: f32,
-    pub amount: f32,
-    pub vol: f32,
+    pub open: f64,
+    pub high: f64,
+    pub low: f64,
+    pub close: f64,
+    pub amount: f64,
+    pub vol: f64,
     pub preclose: f64,
     pub factor: f64,
 }
@@ -45,7 +45,7 @@ impl Day {
                 low: d.low,
                 close: d.close,
                 amount: d.amount,
-                vol: d.vol as f32 / 100.,
+                vol: d.vol as f64 / 100.,
                 preclose: f.preclose,
                 factor: f.factor,
             })
@@ -88,7 +88,7 @@ impl Day {
                 low: d.low,
                 close: d.close,
                 amount: d.amount,
-                vol: d.vol as f32 / 100.,
+                vol: d.vol as f64 / 100.,
                 preclose: f.preclose,
                 factor: f.factor,
             })
